refactor: WebUILottoApplication 기능 Sevice Layer로 분리'
